Refactoring jaxrs-common to rest-common and adding websocket-common.
