+++ commons/src/main/java/org/lfenergy/compas/core/commons/model/package-info.java
@@ -1,0 +1,10 @@
+// SPDX-FileCopyrightText: 2021 Alliander N.V.
+//
+// SPDX-License-Identifier: Apache-2.0
+@XmlSchema(xmlns = {@XmlNs(prefix = "compas-commons", namespaceURI = CommonConstants.COMPAS_COMMONS_V1_NS_URI)})
+package org.lfenergy.compas.core.commons.model;
+
+import org.lfenergy.compas.core.commons.CommonConstants;
+
+import javax.xml.bind.annotation.XmlNs;
+import javax.xml.bind.annotation.XmlSchema;
