+++ commons/src/test/java/org/lfenergy/compas/core/commons/CommonConstantsTest.java
@@ -1,0 +1,15 @@
+// SPDX-FileCopyrightText: 2021 Alliander N.V.
+//
+// SPDX-License-Identifier: Apache-2.0
+package org.lfenergy.compas.core.commons;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertThrows;
+
+class CommonConstantsTest {
+    @Test
+    void constructor_WhenConstructorCalled_ThenShouldThrowExceptionCauseForbidden() {
+        assertThrows(UnsupportedOperationException.class, CommonConstants::new);
+    }
+}
