--- jaxrs-commons/src/test/java/org/lfenergy/compas/core/jaxrs/JaxrsConstantsTest.java
@@ -1,15 +1,0 @@
-// SPDX-FileCopyrightText: 2021 Alliander N.V.
-//
-// SPDX-License-Identifier: Apache-2.0
-package org.lfenergy.compas.core.jaxrs;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.assertThrows;
-
-class JaxrsConstantsTest {
-    @Test
-    void constructor_WhenConstructorCalled_ThenShouldThrowExceptionCauseForbidden() {
-        assertThrows(UnsupportedOperationException.class, JaxrsConstants::new);
-    }
-}
