+++ pom.xml
@@ -44,7 +44,8 @@ SPDX-License-Identifier: Apache-2.0
 
     <modules>
         <module>commons</module>
-        <module>jaxrs-commons</module>
+        <module>rest-commons</module>
+        <module>websocket-commons</module>
         <module>scl-extension</module>
         <module>scl2003</module>
         <module>scl2007b</module>
